chore: rename vuln-mgmt to snapshots and break it down
we will soon want to separate of our logic of somehow acquiring new snapshots and the logic of getting a module's vulnerable functions, in order to more easily refresh the snapshot while the agent is running
