+++ lib/debugger-wrapper.js
@@ -1,7 +1,7 @@
 const debug = require('debug')('snyk:nodejs:inspector');
 const inspector = require('inspector');
 
-const vulnMgmt = require('./vuln-mgmt');
+const snapshot = require('./snapshot');
 const moduleUtils = require('./module-utils');
 const transmitter = require('./transmitter');
 
@@ -24,7 +24,7 @@ function init() {
     throw new Error('Debug session is already connected');
   }
 
-  vulnMgmt.loadsFunctionsToInspect();
+  snapshot.loadsFunctionsToInspect();
 
   session.on('Debugger.scriptParsed', (script) => {
     loadedScripts[script.params.scriptId] = script.params;
@@ -36,7 +36,7 @@ function init() {
     }
 
     const moduleInfo = moduleUtils.getModuleInfo(scriptPath);
-    const vulnerableMethods = vulnMgmt.getVulnerableFunctionsLocations(moduleInfo, scriptPath);
+    const vulnerableMethods = snapshot.getVulnerableFunctionsLocations(moduleInfo, scriptPath);
     Object.keys(vulnerableMethods).forEach((methodName) => {
       const methodLocation = vulnerableMethods[methodName];
       transmitter.addInstrumentedFunction({url: scriptPath, moduleInfo, methodName, methodLocation});
